basic concepts of nan

+++ array properties/trigonometric_equations.py
@@ -2,7 +2,7 @@ import numpy as np
 
 pi = np.pi
 
-mat_one = np.array([0, pi/6, pi/3, pi/2], dtype=np.int8)
+mat_one = np.array([0, pi/6, pi/3, pi/2])
 
 print(np.sin(mat_one))
 print(np.cos(mat_one))
